+++ dflib/src/main/java/org/dflib/FloatSeries.java
@@ -1,5 +1,6 @@
 package org.dflib;
 
+import org.dflib.concat.SeriesConcat;
 import org.dflib.f.FloatPredicate;
 import org.dflib.series.BooleanArraySeries;
 import org.dflib.series.FalseSeries;
@@ -147,7 +148,17 @@ public interface FloatSeries extends Series<Float> {
         return Series.ofFloat(expanded);
     }
 
-    FloatSeries concatFloat(FloatSeries... other);
+    default FloatSeries concatFloat(FloatSeries... other) {
+        if (other.length == 0) {
+            return this;
+        }
+
+        FloatSeries[] combined = new FloatSeries[other.length + 1];
+        combined[0] = this;
+        System.arraycopy(other, 0, combined, 1, other.length);
+
+        return SeriesConcat.floatConcat(combined);
+    }
 
     @Override
     default FloatSeries diff(Series<? extends Float> other) {