+++ dflib/src/main/java/org/dflib/LongSeries.java
@@ -1,5 +1,6 @@
 package org.dflib;
 
+import org.dflib.concat.SeriesConcat;
 import org.dflib.series.BooleanArraySeries;
 import org.dflib.series.FalseSeries;
 import org.dflib.series.LongArraySeries;
@@ -142,7 +143,17 @@ public interface LongSeries extends Series<Long> {
         return Series.ofLong(expanded);
     }
 
-    LongSeries concatLong(LongSeries... other);
+    default LongSeries concatLong(LongSeries... other) {
+        if (other.length == 0) {
+            return this;
+        }
+
+        LongSeries[] combined = new LongSeries[other.length + 1];
+        combined[0] = this;
+        System.arraycopy(other, 0, combined, 1, other.length);
+
+        return SeriesConcat.longConcat(combined);
+    }
 
     @Override
     default LongSeries diff(Series<? extends Long> other) {