+++ dflib/src/main/java/org/dflib/series/FloatBaseSeries.java
@@ -15,7 +15,6 @@ import org.dflib.builder.FloatAccum;
 import org.dflib.builder.IntAccum;
 import org.dflib.builder.ObjectAccum;
 import org.dflib.builder.UniqueFloatAccum;
-import org.dflib.concat.SeriesConcat;
 import org.dflib.f.FloatPredicate;
 import org.dflib.groupby.SeriesGrouper;
 import org.dflib.map.Mapper;
@@ -233,23 +232,6 @@ public abstract class FloatBaseSeries implements FloatSeries {
     public Series<Float> fillNullsForward() {
         // primitive series has no nulls
         return this;
-    }
-
-    @SafeVarargs
-    @Override
-    public final Series<Float> concat(Series<? extends Float>... other) {
-        // concatenating as Float... to concat as FloatSeries, "concatFloat" should be used
-        if (other.length == 0) {
-            return this;
-        }
-
-        // TODO: use SeriesConcat
-
-        Series<Float>[] combined = new Series[other.length + 1];
-        combined[0] = this;
-        System.arraycopy(other, 0, combined, 1, other.length);
-
-        return SeriesConcat.concat(combined);
     }
 
     @Override